requirements.txt
fastapi~=0.115.6
pydantic~=2.10.3
sqlalchemy~=2.0.36
python-dotenv~=1.0.1
pandas~=2.2.3
numpy~=2.2.0
alembic~=1.14.0
asyncpg~=0.30.0
uvicorn~=0.32.1